Add a formal state with a spinner before the pipeline has been prepared
