+++ js_modules/dagit/src/execute/PipelineExecutionContainer.tsx
@@ -1,10 +1,17 @@
 import * as React from "react";
-import gql from "graphql-tag";
-import { Mutation, MutationFn } from "react-apollo";
 import * as yaml from "yaml";
+import gql from "graphql-tag";
+import styled from "styled-components";
+import { Colors, Spinner } from "@blueprintjs/core";
+import { ApolloConsumer, Mutation, MutationFn } from "react-apollo";
+
 import TabBar from "./TabBar";
 import { showCustomAlert } from "../CustomAlertProvider";
-
+import ExecutionStartButton from "./ExecutionStartButton";
+import { RunPreview } from "./RunPreview";
+import { PanelDivider } from "../PanelDivider";
+import SolidSelector from "./SolidSelector";
+import ConfigEditor from "../configeditor/ConfigEditor";
 import {
   applyChangesToSession,
   applySelectSession,
@@ -12,10 +19,20 @@ import {
   applyCreateSession,
   IStorageData,
   IExecutionSession,
-  IExecutionSessionChanges
+  IExecutionSessionChanges,
+  SESSION_CONFIG_PLACEHOLDER
 } from "../LocalStorage";
-import PipelineExecution from "./PipelineExecution";
-import ExecutionStartButton from "./ExecutionStartButton";
+import {
+  CONFIG_EDITOR_PIPELINE_FRAGMENT,
+  CONFIG_EDITOR_VALIDATION_FRAGMENT,
+  scaffoldConfig,
+  responseToValidationResult
+} from "../configeditor/ConfigEditorUtils";
+
+import {
+  PreviewConfigQuery,
+  PreviewConfigQueryVariables
+} from "./types/PreviewConfigQuery";
 import { PipelineExecutionContainerFragment } from "./types/PipelineExecutionContainerFragment";
 import {
   StartPipelineExecution,
@@ -23,52 +40,110 @@ import {
 } from "./types/StartPipelineExecution";
 
 const YAML_SYNTAX_INVALID = `The YAML you provided couldn't be parsed. Please fix the syntax errors and try again.`;
+const CONFIRM_RESET_TO_SCAFFOLD = `Would you like to reset your config to a scaffold based on this subset of the pipeline?`;
 
 interface IPipelineExecutionContainerProps {
-  pipeline?: PipelineExecutionContainerFragment;
-  currentSession: IExecutionSession;
   data: IStorageData;
   onSave: (data: IStorageData) => void;
+  pipeline?: PipelineExecutionContainerFragment;
+  currentSession: IExecutionSession;
+}
+
+interface IPipelineExecutionContainerState {
+  editorVW: number;
+  preview: PreviewConfigQuery | null;
 }
 
 export default class PipelineExecutionContainer extends React.Component<
-  IPipelineExecutionContainerProps
+  IPipelineExecutionContainerProps,
+  IPipelineExecutionContainerState
 > {
   static fragments = {
     PipelineExecutionContainerFragment: gql`
       fragment PipelineExecutionContainerFragment on Pipeline {
         name
-        ...PipelineExecutionPipelineFragment
+        environmentType {
+          key
+        }
+        ...ConfigEditorPipelineFragment
       }
-
-      ${PipelineExecution.fragments.PipelineExecutionPipelineFragment}
+      ${CONFIG_EDITOR_PIPELINE_FRAGMENT}
     `
   };
 
-  handleSelectSession = (session: string) => {
+  state: IPipelineExecutionContainerState = {
+    editorVW: 75,
+    preview: null
+  };
+
+  mounted = false;
+
+  componentDidMount() {
+    this.mounted = true;
+    this.ensureSessionStateValid();
+  }
+
+  componentDidUpdate() {
+    this.ensureSessionStateValid();
+  }
+
+  componentWillUnmount() {
+    this.mounted = false;
+  }
+
+  ensureSessionStateValid() {
+    const { currentSession, pipeline } = this.props;
+    if (!pipeline) return;
+
+    // We have to initialize the sessions in local storage here because the app
+    // needs to have the pieline (with the correct subset) in order to scaffold
+    // the config YAML. In the future this could go in some sort of HOC I suppose.
+    if (currentSession.config === SESSION_CONFIG_PLACEHOLDER) {
+      this.onSaveSession(currentSession.key, {
+        config: scaffoldConfig(pipeline)
+      });
+    }
+  }
+
+  onConfigChange = (config: any) => {
+    this.onSaveSession(this.props.currentSession.key, { config });
+  };
+
+  onSolidSubsetChange = (solidSubset: string[] | null) => {
+    const changes: IExecutionSessionChanges = { solidSubset };
+    if (confirm(CONFIRM_RESET_TO_SCAFFOLD)) {
+      changes.config = SESSION_CONFIG_PLACEHOLDER;
+    }
+
+    this.onSaveSession(this.props.currentSession.key, changes);
+  };
+
+  onSelectSession = (session: string) => {
     this.props.onSave(applySelectSession(this.props.data, session));
   };
 
-  handleSaveSession = (session: string, changes: IExecutionSessionChanges) => {
+  onSaveSession = (session: string, changes: IExecutionSessionChanges) => {
     this.props.onSave(applyChangesToSession(this.props.data, session, changes));
   };
 
-  handleCreateSession = (initial?: IExecutionSessionChanges) => {
+  onCreateSession = (initial?: IExecutionSessionChanges) => {
     this.props.onSave(applyCreateSession(this.props.data, initial));
   };
 
-  handleRemoveSession = (session: string) => {
+  onRemoveSession = (session: string) => {
     this.props.onSave(applyRemoveSession(this.props.data, session));
   };
 
-  handleExecute = async (
+  onExecute = async (
     startPipelineExecution: MutationFn<
       StartPipelineExecution,
       StartPipelineExecutionVariables
     >
   ) => {
     const { pipeline } = this.props;
-    if (!pipeline) {
+    const { preview } = this.state;
+
+    if (!pipeline || !preview) {
       alert(
         "Dagit is still retrieving pipeline info. Please try again in a moment."
       );
@@ -128,33 +203,101 @@ export default class PipelineExecutionContainer extends React.Component<
 
   render() {
     const { currentSession, pipeline } = this.props;
+    const { preview } = this.state;
 
     return (
-      <Mutation<StartPipelineExecution, StartPipelineExecutionVariables>
-        mutation={START_PIPELINE_EXECUTION_MUTATION}
-      >
-        {startPipelineExecution => (
-          <>
+      <>
+        <Mutation<StartPipelineExecution, StartPipelineExecutionVariables>
+          mutation={START_PIPELINE_EXECUTION_MUTATION}
+        >
+          {startPipelineExecution => (
             <TabBar
               sessions={this.props.data.sessions}
               currentSession={currentSession}
-              onSelectSession={this.handleSelectSession}
-              onCreateSession={this.handleCreateSession}
-              onRemoveSession={this.handleRemoveSession}
-              onSaveSession={this.handleSaveSession}
+              onSelectSession={this.onSelectSession}
+              onCreateSession={this.onCreateSession}
+              onRemoveSession={this.onRemoveSession}
+              onSaveSession={this.onSaveSession}
             >
-              <ExecutionStartButton
-                onClick={() => this.handleExecute(startPipelineExecution)}
-              />
+              {!this.state.preview ? (
+                <Spinner size={17} />
+              ) : (
+                <ExecutionStartButton
+                  onClick={() => this.onExecute(startPipelineExecution)}
+                />
+              )}
             </TabBar>
-            <PipelineExecution
-              pipeline={pipeline}
-              currentSession={currentSession}
-              onSaveSession={this.handleSaveSession}
+          )}
+        </Mutation>
+        {currentSession ? (
+          <PipelineExecutionWrapper>
+            <Split width={this.state.editorVW} style={{ flexShrink: 0 }}>
+              <ApolloConsumer>
+                {client => (
+                  <ConfigEditor
+                    readOnly={false}
+                    pipeline={pipeline}
+                    configCode={currentSession.config}
+                    onConfigChange={this.onConfigChange}
+                    checkConfig={async config => {
+                      if (!pipeline) return { isValid: true };
+
+                      const { data } = await client.query<
+                        PreviewConfigQuery,
+                        PreviewConfigQueryVariables
+                      >({
+                        fetchPolicy: "no-cache",
+                        query: PREVIEW_CONFIG_QUERY,
+                        variables: {
+                          config,
+                          pipeline: {
+                            name: pipeline.name,
+                            solidSubset: currentSession.solidSubset
+                          }
+                        }
+                      });
+
+                      this.setState({ preview: data });
+
+                      return responseToValidationResult(
+                        config,
+                        data.isPipelineConfigValid
+                      );
+                    }}
+                  />
+                )}
+              </ApolloConsumer>
+              <SessionSettingsFooter className="bp3-dark">
+                {!pipeline ? (
+                  <span />
+                ) : (
+                  <SolidSelector
+                    pipelineName={pipeline.name}
+                    value={currentSession.solidSubset || null}
+                    onChange={this.onSolidSubsetChange}
+                  />
+                )}
+              </SessionSettingsFooter>
+            </Split>
+            <PanelDivider
+              axis="horizontal"
+              onMove={(vw: number) => this.setState({ editorVW: vw })}
             />
-          </>
+            <Split>
+              {preview ? (
+                <RunPreview
+                  plan={preview.executionPlan}
+                  validation={preview.isPipelineConfigValid}
+                />
+              ) : (
+                <RunPreview />
+              )}
+            </Split>
+          </PipelineExecutionWrapper>
+        ) : (
+          <span />
         )}
-      </Mutation>
+      </>
     );
   }
 }
@@ -182,4 +325,50 @@ const START_PIPELINE_EXECUTION_MUTATION = gql`
       }
     }
   }
+`;
+
+const PREVIEW_CONFIG_QUERY = gql`
+  query PreviewConfigQuery(
+    $pipeline: ExecutionSelector!
+    $config: PipelineConfig!
+  ) {
+    isPipelineConfigValid(pipeline: $pipeline, config: $config) {
+      ...ConfigEditorValidationFragment
+      ...RunPreviewConfigValidationFragment
+    }
+    executionPlan(pipeline: $pipeline, config: $config) {
+      ...RunPreviewExecutionPlanResultFragment
+    }
+  }
+  ${RunPreview.fragments.RunPreviewConfigValidationFragment}
+  ${RunPreview.fragments.RunPreviewExecutionPlanResultFragment}
+  ${CONFIG_EDITOR_VALIDATION_FRAGMENT}
+`;
+
+const PipelineExecutionWrapper = styled.div`
+  flex: 1 1;
+  display: flex;
+  flex-direction: row;
+  width: 100%;
+  height: 100vh;
+  position: absolute;
+  padding-top: 100px;
+`;
+
+const SessionSettingsFooter = styled.div`
+  color: white;
+  display: flex;
+  border-top: 1px solid ${Colors.DARK_GRAY5};
+  background-color: ${Colors.DARK_GRAY2};
+  align-items: center;
+  height: 47px;
+  padding: 8px;
+}
+`;
+
+const Split = styled.div<{ width?: number }>`
+  ${props => (props.width ? `width: ${props.width}vw` : `flex: 1`)};
+  position: relative;
+  flex-direction: column;
+  display: flex;
 `;
